+++ src/backend/model/channel.js
@@ -1,6 +1,6 @@
+require("./category");
 const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
-require("./category");
 
 const channelSchema = new Schema({
   name: {
