+++ src/frontend/app.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-import { App } from "./component/App.jsx";
+import { App } from "./component/app.jsx";
 import "./css/app.css";
 
 const container = document.getElementById("app");
