+++ src/frontend/component/privacy-policy.jsx
